backend/app/requirements.txt
flask
mysql-connector-python
flask-cors
paho-mqtt
